+++ src/App.js
@@ -1,29 +1,13 @@
 import React from "react";
-
-let todoList = [
-  {
-    id: 0,
-    title: "Meditate",
-  },
-  {
-    id: 1,
-    title: "Eat",
-  },
-  {
-    id: 2,
-    title: "Sleep early",
-  },
-];
+import AddTodoForm from "./AddTodoForm";
+import TodoList from "./TodoList";
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function (item) {
-          return <li key={item.id}>{item.title}</li>;
-        })}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
 }
